+++ api/config/config.js
@@ -46,7 +46,9 @@ module.exports = {
     'https://www.googleapis.com/auth/drive.activity.readonly',
     'https://www.googleapis.com/auth/drive.metadata.readonly',
     'https://www.googleapis.com/auth/admin.directory.group',
+    'https://www.googleapis.com/auth/admin.directory.group.readonly',
     'https://www.googleapis.com/auth/admin.reports.audit.readonly',
     'https://www.googleapis.com/auth/apps.groups.settings',
   ],
+  PROJECT_NAME: process.env.PROJECT_NAME,
 }