+++ api/routes/groupsRoutes.js
@@ -2,54 +2,39 @@ const express = require('express')
 const router = express.Router()
 const groupsController = require('../controllers/groupsController')
 const { encryptResponseMiddleware, decryptRequestMiddleware } = require('../controllers/crypto/cryptoMiddleware')
+const { validateJWTMiddleware } = require('../controllers/googleAuthController')
+
+// Global middlewares for all routes
+router.use(validateJWTMiddleware) // Validate JWT for all routes
+router.use(decryptRequestMiddleware) // Decrypt request for all routes
 
 //route to list all groups in customer organization
-router.post('/list', decryptRequestMiddleware, groupsController.listAllGroups, encryptResponseMiddleware)
+router.post('/list', groupsController.listAllGroups)
 
 //route to get group by its email
-router.post('/get', decryptRequestMiddleware, groupsController.getGroup, encryptResponseMiddleware)
+router.post('/get', decryptRequestMiddleware, groupsController.getGroup)
 
 //route to list direct members of a group
-router.post(
-  '/list-direct-members',
-  decryptRequestMiddleware,
-  groupsController.listDirectMembers,
-  encryptResponseMiddleware
-)
+router.post('/list-direct-members', groupsController.listDirectMembers)
 
 //route to list all members of a group(both direct and indirect)
-router.post('/list-all-members', decryptRequestMiddleware, groupsController.listAllMembers, encryptResponseMiddleware)
+router.post('/list-all-members', groupsController.listAllMembers)
 
 //route to get group activity logs(all group logs for all groups in cx domain)
-router.post('/get-activity', decryptRequestMiddleware, groupsController.getGroupActivity, encryptResponseMiddleware)
+router.post('/get-activity', groupsController.getGroupActivity)
 
 //route to get group joined activity(all "add_member" and "accept_invitation" logs for all groups in cx domain)
-router.post(
-  '/get-joined-activity',
-  decryptRequestMiddleware,
-  groupsController.getGroupJoinedActivity,
-  encryptResponseMiddleware
-)
+router.post('/get-joined-activity', groupsController.getGroupJoinedActivity)
 
 //route to get nested membership table for a member(group or user)
 
-router.post(
-  '/get-nested-membership',
-  decryptRequestMiddleware,
-  groupsController.getNestedMembership,
-  encryptResponseMiddleware
-)
+router.post('/get-nested-membership', groupsController.getNestedMembership)
 
 //route to get group hierarchy relative to a group(or potentially in the future a user)
-router.post('/get-hierarchy', decryptRequestMiddleware, groupsController.getGroupHierarchy, encryptResponseMiddleware)
+router.post('/get-hierarchy', groupsController.getGroupHierarchy)
 
 //route to list members of groups in CSV format
-router.post(
-  '/bulk-export',
-  decryptRequestMiddleware,
-  groupsController.listGroupsMembersInExportFormat,
-  encryptResponseMiddleware
-)
+router.post('/bulk-export', groupsController.listGroupsMembersInExportFormat)
 
 //route to update the 'whoCanLeaveGroup' setting of the specified group
 router.put('/update-whocanleave', groupsController.updateWhoCanLeaveGroup)
@@ -59,5 +44,7 @@ router.delete('/delete-members', groupsController.deleteMembers)
 
 //route to delete a member from multiple groups
 router.delete('/delete-member-from-groups', groupsController.deleteMemberFromGroups)
+
+router.use(encryptResponseMiddleware) // Encrypt request for all routes
 
 module.exports = router